Removed cypher-commons
