--- community/cypher/cypher-commons/.gitignore
@@ -1,5 +1,0 @@
-*.iml
-*.ipr
-*.iws
-target
-.DS_Store
