--- community/cypher/cypher-commons/NOTICE.txt
@@ -1,33 +1,0 @@
-Neo4j
-Copyright © 2002-2014 Network Engine for Objects in Lund AB (referred to
-in this notice as "Neo Technology")
-   [http://neotechnology.com]
-
-This product includes software ("Software") developed by Neo Technology.
-
-The copyright in the bundled Neo4j graph database (including the
-Software) is owned by Neo Technology. The Software developed and owned
-by Neo Technology is licensed under the GNU GENERAL PUBLIC LICENSE
-Version 3 (http://www.fsf.org/licensing/licenses/gpl-3.0.html) ("GPL")
-to all third parties and that license, as required by the GPL, is
-included in the LICENSE.txt file.
-
-However, if you have executed an End User Software License and Services
-Agreement or an OEM Software License and Support Services Agreement, or
-another commercial license agreement with Neo Technology or one of its
-affiliates (each, a "Commercial Agreement"), the terms of the license in
-such Commercial Agreement will supersede the GPL and you may use the
-software solely pursuant to the terms of the relevant Commercial
-Agreement.
-
-Full license texts are found in LICENSES.txt.
-
-Third-party licenses
---------------------
-
-Apache Software License, Version 2.0
-  ConcurrentLinkedHashMap
-
-BSD - Scala License
-  Scala Library
-
