+++ community/cypher/cypher-compiler-2.2/LICENSES.txt
@@ -3,6 +3,7 @@ libraries. For an overview of the licenses see the NOTICE.txt file.
 
 ------------------------------------------------------------------------------
 Apache Software License, Version 2.0
+  ConcurrentLinkedHashMap
   opencsv
   parboiled-core
   parboiled-scala
