+++ community/cypher/cypher-compiler-2.2/NOTICE.txt
@@ -26,6 +26,7 @@ Third-party licenses
 --------------------
 
 Apache Software License, Version 2.0
+  ConcurrentLinkedHashMap
   opencsv
   parboiled-core
   parboiled-scala
