+++ community/cypher/cypher-compiler-2.2/src/main/java/org/neo4j/cypher/javacompat/PlanDescription.java
@@ -19,10 +19,12 @@
  */
 package org.neo4j.cypher.javacompat;
 
+
+import org.neo4j.cypher.ProfilerStatisticsNotReadyException;
+
 import java.util.List;
 import java.util.Map;
 
-import org.neo4j.cypher.ProfilerStatisticsNotReadyException;
 
 /**
  * Instances describe single execution steps in a Cypher query execution plan
