+++ community/cypher/cypher-compiler-2.2/src/main/scala/org/neo4j/cypher/internal/helpers/DynamicIterable.scala
@@ -17,16 +17,25 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.neo4j.cypher
+package org.neo4j.cypher.internal.helpers
 
-import javacompat.{PlanDescription => JPlanDescription}
+import java.lang.{Iterable => JIterable}
+import java.util.{Iterator => JIterator}
 
-/**
- * Abstract description of an execution plan
- */
-trait PlanDescription {
-  def name: String
-  def asJava: JPlanDescription
-  def render(builder: StringBuilder, separator: String, levelSuffix: String)
-  def render(builder: StringBuilder)
+import scala.collection.JavaConverters._
+
+object DynamicIterable {
+  def apply[T](f: => Iterator[T]) = new Iterable[T] {
+    override def iterator = f
+  }
+}
+
+object DynamicJavaIterable {
+  def apply[T](f: => Iterator[T]) = new JIterable[T] {
+    override def iterator = f.asJava
+  }
+
+  def apply[T](iterable: Iterable[T]) = new JIterable[T] {
+    override def iterator: JIterator[T] = iterable.iterator.asJava
+  }
 }
