+++ community/cypher/cypher-compiler-2.2/src/main/scala/org/neo4j/cypher/internal/helpers/PartialFunctionSupport.scala
