+++ community/cypher/cypher-compiler-2.2/src/main/scala/org/neo4j/cypher/internal/helpers/StringRenderingSupport.scala
@@ -33,4 +33,3 @@ trait StringRenderingSupport {
     builder ++= super.toString
   }
 }
-
