+++ community/cypher/cypher-compiler-2.2/src/test/scala/org/neo4j/cypher/internal/commons/CypherFunSuite.scala
