+++ community/cypher/docs/cypher-docs/pom.xml
@@ -84,13 +84,6 @@
       <version>${project.version}</version>
       <scope>test</scope>
     </dependency>
-    <dependency>
-      <groupId>org.neo4j</groupId>
-      <artifactId>neo4j-cypher-commons</artifactId>
-      <version>${project.version}</version>
-      <scope>provided</scope>
-      <type>test-jar</type>
-    </dependency>
 
     <!-- other -->
