+++ community/neo4j-community/pom.xml
@@ -122,10 +122,6 @@ the relevant Commercial Agreement.
                   <filtering>false</filtering>
                 </resource>
                 <resource>
-                  <directory>${project.build.directory}/../../cypher/cypher-commons/src/main/java/</directory>
-                  <filtering>false</filtering>
-                </resource>
-                <resource>
                   <directory>${project.build.directory}/../../graph-algo/src/main/java/</directory>
                   <filtering>false</filtering>
                 </resource>
